CEC/requirements.txt
Django==5.0.0
google-generativeai==0.3.2
